+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsBoxTestGenerated.java
@@ -508,6 +508,12 @@ public class JsBoxTestGenerated extends AbstractJsBoxTest {
       }
 
       @Test
+      @TestMetadata("generatedFactories.kt")
+      public void testGeneratedFactories() {
+        run("generatedFactories.kt");
+      }
+
+      @Test
       @TestMetadata("implTypeNotDirectlyAccessible.kt")
       public void testImplTypeNotDirectlyAccessible() {
         run("implTypeNotDirectlyAccessible.kt");
@@ -547,6 +553,12 @@ public class JsBoxTestGenerated extends AbstractJsBoxTest {
       @TestMetadata("intoSet.kt")
       public void testIntoSet() {
         run("intoSet.kt");
+      }
+
+      @Test
+      @TestMetadata("memberInjectionAcrossModules.kt")
+      public void testMemberInjectionAcrossModules() {
+        run("memberInjectionAcrossModules.kt");
       }
 
       @Test